C B C A D C B A C D B C B D B D A B A

+++ wolf3dredux_0.01/game/wolf3d/entities/wolf_bj.c
@@ -44,17 +44,19 @@
 
 -----------------------------------------------------------------------------
 */
-void SpawnBJVictory( void )
+void SpawnBJVictory(void)
 {
 	entity_t *bj;
 
-	bj = SpawnActor( en_bj, POS2TILE(Player.position.origin[0]), POS2TILE(Player.position.origin[1]), dir4_north, r_world );
-	if( ! bj ) {
+	bj = SpawnActor(en_bj, (int)POS2TILE(Player.position.origin[0]),
+					(int)POS2TILE(Player.position.origin[1]), dir4_north,
+					r_world);
+	if (! bj) {
 		return;
 	}
 
-	bj->x = Player.position.origin[ 0 ];
-	bj->y = Player.position.origin[ 1 ];
+	bj->x = (int)Player.position.origin[0];
+	bj->y = (int)Player.position.origin[1];
 	bj->state = st_path1;
 	bj->speed = BJRUNSPEED;
 	bj->flags = FL_NONMARK; /* FL_NEVERMARK; */
@@ -101,10 +103,10 @@ void T_BJRun( entity_t *Guard )
 
 -----------------------------------------------------------------------------
 */
-void T_BJJump( entity_t *Guard )
+void T_BJJump(entity_t *Guard)
 {
 #if 0
-	MoveObj(Guard, Guard->speed);
+	MoveObj(Guard, Guard->speed); /* MoveObj is not yet implemented */
 #endif /* 0 */
 }
 
@@ -120,9 +122,11 @@ void T_BJJump( entity_t *Guard )
 
 -----------------------------------------------------------------------------
 */
-void T_BJYell( entity_t *Guard )
+void T_BJYell(entity_t *Guard)
 {
-	Sound_StartSound( NULL, 0, CHAN_VOICE, Sound_RegisterSound( "sfx/082.wav" ), 1, ATTN_NORM, 0 );
+	/* TODO: use parameter 'Guard' */
+	Sound_StartSound(NULL, 0, CHAN_VOICE, Sound_RegisterSound("sfx/082.wav"),
+					 1, ATTN_NORM, 0);
 }
 
 /*
@@ -137,8 +141,9 @@ void T_BJYell( entity_t *Guard )
 
 -----------------------------------------------------------------------------
 */
-void T_BJDone( entity_t *Guard )
+void T_BJDone(entity_t *Guard)
 {
+	/* TODO: use parameter 'Guard' */
 	Player.playstate = ex_victory; /* exit castle tile */
 }
